api/admin/requirements.txt
flask
flask-restful
psycopg2-binary